refactor(web): promote SpeakerTranscriptDisplay + dotColorForLabel to lib
Both shorts-editor (existing) and shorts-auto (incoming) need to render
speaker-diarized transcript turns. Lifting the inline implementation
out of features/shorts-editor/components/SceneListPanel.tsx into a
shared lib module so the new auto-shorts inspector script panel can
consume it without crossing feature boundaries.

The shared component gains a ``className`` prop so the editor's
existing ``mb-1.5`` spacing is preserved at the call site instead of
being baked into the component. Behavior of the editor's scene list
is unchanged (covered by existing tests).

Also syncs lib/types/shorts-auto.ts ClipMemberResponse with the
backend's PR 1 additions: optional ``transcript`` and
``scene_caption``, plus removal of the unused ``reason?: string | null``
field that drifted onto the frontend type but was never emitted by the
server (flagged during pre-merge multi-reviewer pass).

Test for ``dotColorForLabel`` moves to ``lib/__tests__/`` next to its
new home so the editor feature no longer owns it.

Constraint: no cross-feature imports; both consumers reach through lib/
Rejected: re-exporting dotColorForLabel from SceneListPanel.tsx | leaves an unrelated feature owning a shared util
Confidence: high
Scope-risk: narrow
